+++ changelog.md
@@ -2,7 +2,9 @@
 
 ### 1.1.0, soon...
 
+* fixed: support for WooCommerce HPOS (high performance order storage)
 * added: filter hook `mininum_age_woo_show_question` allowing hookers to conditionally suppress the age question
+* changed: marked as tested up to WooCommerce 8.2
 
 ### 1.0.18, 2022-11-22